add page_utils file to handle pages
