+++ app/main/routes.py
@@ -41,6 +41,10 @@ from app.main.util.filter_sort_builder import (
     build_filters,
     build_sorting_orders,
 )
+from app.main.util.page_utils import (
+    get_page_and_per_page,
+    redirect_if_page_invalid,
+)
 from app.main.util.pagination import (
     calculate_total_pages,
     get_pagination,
@@ -166,26 +170,6 @@ def callback():
 @bp.route("/accessibility", methods=["GET"])
 def accessibility():
     return render_template("accessibility.html")
-
-
-def get_page_and_per_page(validated_data):
-    page = validated_data.get("page")
-    per_page = validated_data.get("per_page")
-    if not per_page:
-        per_page = int(current_app.config["DEFAULT_PAGE_SIZE"])
-    return page, per_page
-
-
-def redirect_if_page_invalid(requested_page, default_page, endpoint, **kwargs):
-    """
-    Redirects to the default page if the requested page is invalid (e.g., out of range).
-    Returns a redirect response if a redirect is needed, otherwise None.
-    """
-    if requested_page != default_page:
-        # Copy current query parameters and set page to default_page
-        args = request.args.to_dict()
-        args["page"] = default_page
-        return redirect(url_for(endpoint, **kwargs, **args))
 
 
 @bp.route("/browse", methods=["GET"])
